requirements.txt
django==5.0.6
django-ninja==1.1.0
django-ninja-jwt==5.3.1
hydroserver-sensorthings==0.3.0
dj-database-url==2.2.0
python-decouple==3.8
pytz==2024.1
pyyaml==6.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
django-cors-headers==4.4.0
urllib3==2.2.2
boto3==1.34.131
authlib==1.3.1
django-storages==1.14.3
django-simple-history==3.7.0
hsclient==1.0.1
hsmodels==1.0.3
country-list==1.1.0
